Add null check lambda.

+++ dispatch-service/dispatch-service/src/main/java/intelligent_taxi/dispatchservice/command/DispatchCommandService.java
@@ -3,10 +3,13 @@ package intelligent_taxi.dispatchservice.command;
 import intelligent_taxi.dispatchservice.clientWrapper.MemberClientWrapper;
 import intelligent_taxi.dispatchservice.clientWrapper.PayClientWrapper;
 import intelligent_taxi.dispatchservice.clientWrapper.TaxiClientWrapper;
+import intelligent_taxi.dispatchservice.controller.restResponse.ResponseMessage;
 import intelligent_taxi.dispatchservice.domain.Dispatch;
+import intelligent_taxi.dispatchservice.domain.DispatchState;
 import intelligent_taxi.dispatchservice.dto.dispatch.DispatchRequest;
 import intelligent_taxi.dispatchservice.dto.dispatch.RequestDispatch;
 import intelligent_taxi.dispatchservice.dto.taxi.TaxiResponse;
+import intelligent_taxi.dispatchservice.exception.DispatchCustomException;
 import intelligent_taxi.dispatchservice.repository.DispatchRepository;
 import intelligent_taxi.dispatchservice.validator.ServiceValidator;
 import lombok.RequiredArgsConstructor;
@@ -36,17 +39,31 @@ public class DispatchCommandService {
     }
 
     public void removeDispatch(Long id, String username) {
-        Dispatch dispatch = serviceValidator.validateRemoveDispatch(id, username);
+        Dispatch dispatch = dispatchRepository.findOneById(id)
+                .orElseThrow(() -> new DispatchCustomException(ResponseMessage.DISPATCH_IS_NULL));
+
+        if (!username.equals(dispatch.getUsername())) {
+            throw new DispatchCustomException(ResponseMessage.NOT_MATCH_USERNAME);
+        }
+
+        if (dispatch.getDispatchState() == DispatchState.FINISH) {
+            throw new DispatchCustomException(ResponseMessage.DISPATCH_IS_ALREADY_FINISH);
+        }
+
         dispatchRepository.delete(dispatch);
     }
 
     public void rollbackDispatch(Long id) {
-        Dispatch dispatch = dispatchRepository.findOneById(id);
+        Dispatch dispatch = dispatchRepository.findOneById(id)
+                .orElseThrow(() -> new DispatchCustomException(ResponseMessage.DISPATCH_IS_NULL));
         dispatchRepository.delete(dispatch);
     }
 
     public void dispatch(RequestDispatch requestDto, String username) {
-        Dispatch dispatch = serviceValidator.validateDispatch(requestDto);
+        Dispatch dispatch = dispatchRepository.findOneWithinDistance(
+                requestDto.getPresentLatitude(),
+                requestDto.getPresentLongitude()
+        ).orElseThrow(() -> new DispatchCustomException(ResponseMessage.DISPATCH_IS_NULL));
         dispatch.finishDispatch();
 
         TaxiResponse taxi = taxiClientWrapper.getTaxiInfoByUsername(username);
